--- client/app/js/http/getCurrentUser.js
@@ -1,5 +1,0 @@
-import httpGet from './httpGet';
-
-export default function getCurrentUser(token) {
-  return httpGet({ url: '/api/current_user', token });
-}
