--- client/app/js/http/getLeagueFantasyTeams.js
@@ -1,8 +1,0 @@
-import httpGet from './httpGet';
-
-export default function getLeagueFantasyTeams(fantasyLeagueId, token) {
-  return httpGet({
-    url: `/dev_api/league/${fantasyLeagueId}/fantasy_teams/`,
-    token
-  });
-}
