--- client/app/js/http/getToken.js
@@ -1,5 +1,0 @@
-import httpGet from './httpGet';
-
-export default function getToken(email, password) {
-  return httpGet({ url: '/api/token', auth: { u: email, p: password } });
-}
